Various fixes
mp-ibc-join:
- Fix the handling of boundary QShift when adding sites to the window.

mp-ibc-apply:
- Allow the QShift of the right boundary to be specified.
- Add more error checking for incompatible boundaries.

+++ common/rangelist.cpp
@@ -73,6 +73,6 @@ RangeList::RangeList(std::string Str)
    List = std::vector<double>();
    for (int i = 0; i < Num; ++i)
    {
-      List.push_back(Start * (Num-1-i) / (Num-1) + End * i / (Num-1));
+      List.push_back(Start * ((double) (Num-1-i) / (Num-1)) + End * ((double) i / (Num-1)));
    }
 }
